Update results_verification.R
fixed some bugs with saving template_DT to environment (still might be wonky).

Added new function to page through entire recording and create verification annotations. Did this to create gold standards for ML models that used a certain temp_length to assess performance, but thought it could be handy for anyone wanting to move through a recording like this.